Refactored into a Library namespace. Closes #28.
